test(functional): prove lifecycle listener cleanup
