+++ tests/functional/workers/transports/cli/run/lifecycle/lifecycle_adverse_test.go
@@ -9,8 +9,6 @@ import (
 	"io"
 	"net/http"
 	"net/url"
-	"os"
-	"os/exec"
 	"path/filepath"
 	"strings"
 	"sync"
@@ -38,21 +36,23 @@ const lifecyclePartialCodexStdout = "" +
 	`{"type":"item.completed","item":{"id":"partial-codex-message","type":"agent_message","text":"partial provider output"}}` + "\n"
 
 type hostedLifecycleInvocation struct {
-	coordinator  *lifecycleCoordinator
-	api          *support.ProcessAPIServer
-	shutdownGate *lifecycleGate
-	command      *support.ProcessCommand
-	inputs       *support.CapturedInputs
-	baseURL      string
+	coordinator   *lifecycleCoordinator
+	api           *support.ProcessAPIServer
+	shutdownGate  *lifecycleGate
+	command       *support.ProcessCommand
+	inputs        *support.CapturedInputs
+	baseURL       string
+	listenerClose <-chan struct{}
 }
 
 type hostedCancelableLifecycleInvocation struct {
-	coordinator  *lifecycleCoordinator
-	api          *support.ProcessAPIServer
-	shutdownGate *lifecycleGate
-	command      *lifecycleCancelableCommand
-	inputs       *support.CapturedInputs
-	baseURL      string
+	coordinator   *lifecycleCoordinator
+	api           *support.ProcessAPIServer
+	shutdownGate  *lifecycleGate
+	command       *lifecycleCancelableCommand
+	inputs        *support.CapturedInputs
+	baseURL       string
+	listenerClose <-chan struct{}
 }
 
 type lifecycleWorkStatusObservation struct {
@@ -79,7 +79,7 @@ func startHostedLifecycleInvocation(
 	prompt string,
 ) *hostedLifecycleInvocation {
 	t.Helper()
-	api := support.NewProcessAPIServer()
+	api := newLifecycleAPIServer()
 	shutdownGate := newLifecycleGate("adverse listener shutdown")
 	api.HoldShutdownUntilSignaled(shutdownGate.channel())
 	coordinator := buildLifecycleProcess(t, serviceedges.Edges{
@@ -97,17 +97,18 @@ func startHostedLifecycleInvocation(
 		prompt,
 	}, factoryDir)
 	command := coordinator.StartCommand(inputs)
-	baseURL, err := coordinator.WaitForReadiness(api)
+	baseURL, err := coordinator.WaitForReadiness(api.server)
 	if err != nil {
 		t.Fatal(err)
 	}
 	return &hostedLifecycleInvocation{
-		coordinator:  coordinator,
-		api:          api,
-		shutdownGate: shutdownGate,
-		command:      command,
-		inputs:       inputs,
-		baseURL:      baseURL,
+		coordinator:   coordinator,
+		api:           api.server,
+		shutdownGate:  shutdownGate,
+		command:       command,
+		inputs:        inputs,
+		baseURL:       baseURL,
+		listenerClose: api.closed,
 	}
 }
 
@@ -118,7 +119,7 @@ func startHostedCancelableLifecycleInvocation(
 	prompt string,
 ) *hostedCancelableLifecycleInvocation {
 	t.Helper()
-	api := support.NewProcessAPIServer()
+	api := newLifecycleAPIServer()
 	shutdownGate := newLifecycleGate("cancelable listener shutdown")
 	api.HoldShutdownUntilSignaled(shutdownGate.channel())
 	coordinator := buildLifecycleProcess(t, serviceedges.Edges{
@@ -136,17 +137,18 @@ func startHostedCancelableLifecycleInvocation(
 		prompt,
 	}, factoryDir)
 	command := coordinator.StartCancelableCommand(inputs)
-	baseURL, err := coordinator.WaitForReadiness(api)
+	baseURL, err := coordinator.WaitForReadiness(api.server)
 	if err != nil {
 		t.Fatal(err)
 	}
 	return &hostedCancelableLifecycleInvocation{
-		coordinator:  coordinator,
-		api:          api,
-		shutdownGate: shutdownGate,
-		command:      command,
-		inputs:       inputs,
-		baseURL:      baseURL,
+		coordinator:   coordinator,
+		api:           api.server,
+		shutdownGate:  shutdownGate,
+		command:       command,
+		inputs:        inputs,
+		baseURL:       baseURL,
+		listenerClose: api.closed,
 	}
 }
 
@@ -184,7 +186,7 @@ func runLifecyclePartialProviderOutput(t *testing.T) {
 		t.Fatal("partial provider Process.Execute error = nil, want terminal failure")
 	}
 	assertLifecycleFailureOutput(t, invocation.inputs, err)
-	finishHostedLifecycleInvocation(t, invocation.coordinator, invocation.baseURL, workID, workerSessionID)
+	finishHostedLifecycleInvocation(t, invocation.coordinator, invocation.baseURL, invocation.listenerClose, workID, workerSessionID)
 }
 
 func runLifecycleServerAttachedProviderFailure(t *testing.T) {
@@ -219,7 +221,7 @@ func runLifecycleServerAttachedProviderFailure(t *testing.T) {
 		t.Fatal("server-attached failure Process.Execute error = nil, want terminal failure")
 	}
 	assertLifecycleFailureOutput(t, invocation.inputs, err)
-	finishHostedLifecycleInvocation(t, invocation.coordinator, invocation.baseURL, workID, workerSessionID)
+	finishHostedLifecycleInvocation(t, invocation.coordinator, invocation.baseURL, invocation.listenerClose, workID, workerSessionID)
 }
 
 func runLifecycleCancellationAndRecovery(t *testing.T) {
@@ -285,7 +287,7 @@ func runLifecycleCancellationAndRecovery(t *testing.T) {
 	if !strings.Contains(strings.ToLower(err.Error()), "cancel") {
 		t.Fatalf("canceled Process.Execute error = %v, want cancellation-compatible diagnostic", err)
 	}
-	finishCancelableHostedLifecycleInvocation(t, invocation.coordinator, invocation.baseURL, workID, workerSessionID)
+	finishCancelableHostedLifecycleInvocation(t, invocation.coordinator, invocation.baseURL, invocation.listenerClose, workID, workerSessionID)
 
 	recoveryWorkID, recoveryWorkerSessionID := runHostedLifecycleRecovery(t, factoryDir)
 	if recoveryWorkID == workID {
@@ -299,7 +301,7 @@ func runLifecycleCancellationAndRecovery(t *testing.T) {
 func runLifecycleObservationTimeout(t *testing.T) {
 	factoryDir := scaffoldProviderBackedFactory(t)
 	runner := newBlockingLifecycleRunner()
-	api := support.NewProcessAPIServer()
+	api := newLifecycleAPIServer()
 	shutdownGate := newLifecycleGate("observation-timeout listener shutdown")
 	api.HoldShutdownUntilSignaled(shutdownGate.channel())
 	coordinator := buildLifecycleProcess(t, serviceedges.Edges{
@@ -316,7 +318,7 @@ func runLifecycleObservationTimeout(t *testing.T) {
 		"keep the observation target absent until the local deadline fires",
 	}, factoryDir)
 	command := coordinator.StartCancelableCommand(inputs)
-	baseURL, err := coordinator.WaitForReadiness(api)
+	baseURL, err := coordinator.WaitForReadiness(api.server)
 	if err != nil {
 		t.Fatal(err)
 	}
@@ -342,7 +344,7 @@ func runLifecycleObservationTimeout(t *testing.T) {
 		t.Fatalf("observation-timeout stdout = %q, want no success result", inputs.Stdout())
 	}
 	closeLifecycleCoordinator(t, coordinator)
-	if !lifecycleListenerClosed(baseURL) {
+	if !lifecycleListenerClosed(baseURL, api.closed) {
 		t.Fatalf("observation-timeout listener at %s remained reachable after cancellation and close", baseURL)
 	}
 }
@@ -593,7 +595,7 @@ func runHostedLifecycleRecovery(t *testing.T, factoryDir string) (string, string
 	if invocation.inputs.Stderr() != "" {
 		t.Fatalf("recovery stderr = %q, want empty", invocation.inputs.Stderr())
 	}
-	finishHostedLifecycleInvocation(t, invocation.coordinator, invocation.baseURL, workID, worker.WorkerSessionId)
+	finishHostedLifecycleInvocation(t, invocation.coordinator, invocation.baseURL, invocation.listenerClose, workID, worker.WorkerSessionId)
 	return workID, worker.WorkerSessionId
 }
 
@@ -669,7 +671,9 @@ func waitForLifecycleTerminalWorkText(baseURL, wantText string) (string, bool, s
 func finishHostedLifecycleInvocation(
 	t *testing.T,
 	coordinator *lifecycleCoordinator,
-	baseURL, workID, workerSessionID string,
+	baseURL string,
+	listenerClose <-chan struct{},
+	workID, workerSessionID string,
 ) {
 	t.Helper()
 	if coordinator == nil {
@@ -679,7 +683,8 @@ func finishHostedLifecycleInvocation(
 		t.Fatalf("hosted lifecycle unreleased gates before close = %v", unreleased)
 	}
 	closeLifecycleCoordinator(t, coordinator)
-	if !lifecycleListenerClosed(baseURL) {
+	waitLifecycleSignal(t, listenerClose, "hosted lifecycle listener close")
+	if !lifecycleListenerClosed(baseURL, listenerClose) {
 		t.Fatalf("hosted lifecycle listener at %s remained reachable after Work %q / Worker Session %q close", baseURL, workID, workerSessionID)
 	}
 }
@@ -687,14 +692,17 @@ func finishHostedLifecycleInvocation(
 func finishCancelableHostedLifecycleInvocation(
 	t *testing.T,
 	coordinator *lifecycleCoordinator,
-	baseURL, workID, workerSessionID string,
+	baseURL string,
+	listenerClose <-chan struct{},
+	workID, workerSessionID string,
 ) {
 	t.Helper()
 	if unreleased := coordinator.unreleasedGates(); len(unreleased) != 0 {
 		t.Fatalf("cancelable hosted lifecycle unreleased gates before close = %v", unreleased)
 	}
 	closeLifecycleCoordinator(t, coordinator)
-	if !lifecycleListenerClosed(baseURL) {
+	waitLifecycleSignal(t, listenerClose, "cancelable hosted lifecycle listener close")
+	if !lifecycleListenerClosed(baseURL, listenerClose) {
 		t.Fatalf("cancelable hosted lifecycle listener at %s remained reachable after Work %q / Worker Session %q close", baseURL, workID, workerSessionID)
 	}
 }
@@ -744,23 +752,6 @@ func readLifecycleHTTPJSON[T any](endpoint string) (T, error) {
 		return result, fmt.Errorf("decode GET %s: %w", endpoint, err)
 	}
 	return result, nil
-}
-
-func lifecycleListenerClosed(baseURL string) bool {
-	if strings.TrimSpace(baseURL) == "" {
-		return false
-	}
-	ctx, cancel := context.WithTimeout(context.Background(), lifecycleHTTPTimeout)
-	defer cancel()
-	request, err := http.NewRequestWithContext(ctx, http.MethodGet, strings.TrimSuffix(baseURL, "/")+"/status", nil)
-	if err != nil {
-		return false
-	}
-	response, err := lifecycleHTTPClient.Do(request)
-	if response != nil {
-		response.Body.Close()
-	}
-	return err != nil
 }
 
 func waitLifecycleSignal(t *testing.T, signal <-chan struct{}, name string) {
@@ -862,181 +853,6 @@ func waitCancelableLifecycleCommand(command *lifecycleCancelableCommand) error {
 	case <-timer.C:
 		return fmt.Errorf("cancelable lifecycle command did not complete within %s", lifecycleCommandDoneTimeout)
 	}
-}
-
-type forcedLifecycleCleanupReport struct {
-	ApplicationPID    int    `json:"application_pid"`
-	ProcessClosed     bool   `json:"process_closed"`
-	ProcessCloseError string `json:"process_close_error,omitempty"`
-	ProcessCloseTime  string `json:"process_close_time"`
-	CommandDone       bool   `json:"command_done"`
-	ProviderCalls     int    `json:"provider_calls"`
-	ProviderStarted   bool   `json:"provider_started"`
-	ProviderFinished  bool   `json:"provider_finished"`
-	ProviderCanceled  bool   `json:"provider_canceled"`
-	GatesReleased     bool   `json:"gates_released"`
-	ListenerClosed    bool   `json:"listener_closed"`
-	SessionObserved   bool   `json:"session_observed"`
-	FactoryAbsent     bool   `json:"factory_absent"`
-	ArtifactAbsent    bool   `json:"artifact_absent"`
-}
-
-func runForcedLifecycleCleanupParent(t *testing.T) {
-	t.Helper()
-	reportPath := filepath.Join(t.TempDir(), "forced-worker-cli-lifecycle-cleanup.json")
-	command := exec.Command(os.Args[0], "-test.run=^TestCLIRunCleanInvocationFailurePreservesPublicError$")
-	command.Env = append(os.Environ(),
-		lifecycleForcedCleanupChildEnv+"=1",
-		lifecycleForcedCleanupReportEnv+"="+reportPath,
-	)
-	output, err := command.CombinedOutput()
-	if err == nil {
-		t.Fatalf("forced lifecycle cleanup child exited successfully; output=%q", output)
-	}
-	if command.Process == nil || command.ProcessState == nil || !command.ProcessState.Exited() {
-		t.Fatalf("forced lifecycle cleanup child did not exit; error=%v output=%q", err, output)
-	}
-	if command.ProcessState.ExitCode() == 0 {
-		t.Fatalf("forced lifecycle cleanup child exit code = 0; output=%q", output)
-	}
-	payload, err := os.ReadFile(reportPath)
-	if err != nil {
-		t.Fatalf("read forced lifecycle cleanup report %q: %v; child output=%q", reportPath, err, output)
-	}
-	var report forcedLifecycleCleanupReport
-	if err := json.Unmarshal(payload, &report); err != nil {
-		t.Fatalf("decode forced lifecycle cleanup report %q: %v; child output=%q", reportPath, err, output)
-	}
-	if report.ApplicationPID != command.Process.Pid {
-		t.Fatalf("forced lifecycle application PID = %d, want child PID %d", report.ApplicationPID, command.Process.Pid)
-	}
-	if !report.ProcessClosed || report.ProcessCloseError != "" || report.ProcessCloseTime == "" {
-		t.Fatalf("forced lifecycle process close = %#v, want clean close", report)
-	}
-	if !report.CommandDone || report.ProviderCalls != 1 || !report.ProviderStarted || !report.ProviderFinished || !report.ProviderCanceled {
-		t.Fatalf("forced lifecycle command/provider state = %#v, want one joined canceled command", report)
-	}
-	if !report.GatesReleased || !report.ListenerClosed || !report.SessionObserved {
-		t.Fatalf("forced lifecycle public/resource state = %#v, want released gates, closed listener, observed session", report)
-	}
-	if !report.FactoryAbsent || !report.ArtifactAbsent {
-		t.Fatalf("forced lifecycle owned paths remain: %#v", report)
-	}
-}
-
-func runForcedLifecycleCleanupChild(t *testing.T) {
-	t.Helper()
-	reportPath := strings.TrimSpace(os.Getenv(lifecycleForcedCleanupReportEnv))
-	if reportPath == "" {
-		t.Fatal("forced lifecycle cleanup report path is required")
-	}
-	var coordinator *lifecycleCoordinator
-	var command *lifecycleCancelableCommand
-	var runner *blockingLifecycleRunner
-	var baseURL string
-	var factoryDir string
-	var artifactPath string
-	var sessionObserved atomic.Bool
-	// Register this before every acquired resource so the report is written
-	// after coordinator, command, artifact, and factory cleanup has run.
-	t.Cleanup(func() {
-		report := forcedLifecycleCleanupReport{ApplicationPID: os.Getpid()}
-		if coordinator != nil {
-			report.ProcessClosed, report.ProcessCloseTime = forcedLifecycleProcessClosed(coordinator)
-			if closeErr, _ := coordinator.closeResult(); closeErr != nil {
-				report.ProcessCloseError = closeErr.Error()
-			}
-			report.GatesReleased = len(coordinator.unreleasedGates()) == 0
-		}
-		if command != nil {
-			select {
-			case <-command.Done():
-				report.CommandDone = true
-			default:
-			}
-		}
-		if runner != nil {
-			report.ProviderCalls = int(runner.calls.Load())
-			report.ProviderStarted = isClosed(runner.started)
-			report.ProviderFinished = isClosed(runner.finished)
-			report.ProviderCanceled = runner.CancellationCount() == 1
-		}
-		report.ListenerClosed = lifecycleListenerClosed(baseURL)
-		report.SessionObserved = sessionObserved.Load()
-		report.FactoryAbsent = pathAbsent(factoryDir)
-		report.ArtifactAbsent = pathAbsent(artifactPath)
-		payload, err := json.MarshalIndent(report, "", "  ")
-		if err == nil {
-			err = os.WriteFile(reportPath, payload, 0o600)
-		}
-		if err != nil {
-			t.Errorf("write forced lifecycle cleanup report: %v", err)
-		}
-	})
-
-	var err error
-	factoryDir = scaffoldProviderBackedFactory(t)
-	artifactPath = filepath.Join(factoryDir, "forced-cleanup-artifact.txt")
-	if err := os.WriteFile(artifactPath, []byte("owned lifecycle artifact"), 0o600); err != nil {
-		t.Fatal(err)
-	}
-	t.Cleanup(func() { _ = os.Remove(artifactPath) })
-
-	runner = newBlockingLifecycleRunner()
-	api := support.NewProcessAPIServer()
-	shutdownGate := newLifecycleGate("forced cleanup listener shutdown")
-	api.HoldShutdownUntilSignaled(shutdownGate.channel())
-	coordinator = buildLifecycleProcess(t, serviceedges.Edges{
-		APIServerStarter:      api.Start,
-		ProviderCommandRunner: runner,
-	})
-	coordinator.TrackGate(shutdownGate)
-	inputs := coordinator.Inputs([]string{
-		"you", "run",
-		"--factory", filepath.Join(factoryDir, "factory.json"),
-		"--with-server",
-		"--no-record",
-		"--quiet",
-		"force lifecycle cleanup after acquisition",
-	}, factoryDir)
-	command = coordinator.StartCancelableCommand(inputs)
-	baseURL, err = coordinator.WaitForReadiness(api)
-	if err != nil {
-		t.Fatal(err)
-	}
-	waitLifecycleSignal(t, runner.started, "forced-cleanup provider start")
-	_, err = support.WaitForObservation(
-		lifecycleAdverseSignalTimeout,
-		func() (factoryapi.FactorySession, error) {
-			session, ok, diagnostic := tryReadDefaultFactorySession(baseURL)
-			if !ok {
-				return factoryapi.FactorySession{}, errors.New(diagnostic)
-			}
-			return session, nil
-		},
-		func(session factoryapi.FactorySession) bool { return strings.TrimSpace(session.Id) != "" },
-	)
-	if err != nil {
-		t.Fatal(err)
-	}
-	sessionObserved.Store(true)
-	t.Fatal("intentional lifecycle assertion failure after process, session, listener, provider, gate, and artifact acquisition")
-}
-
-func forcedLifecycleProcessClosed(coordinator *lifecycleCoordinator) (bool, string) {
-	if coordinator == nil {
-		return false, ""
-	}
-	_, duration := coordinator.closeResult()
-	return coordinator.closed(), duration.String()
-}
-
-func pathAbsent(path string) bool {
-	if strings.TrimSpace(path) == "" {
-		return false
-	}
-	_, err := os.Stat(path)
-	return errors.Is(err, os.ErrNotExist)
 }
 
 func isClosed(signal <-chan struct{}) bool {
